zephyr/tests: Add pseudo-unit tests for usocket module.
These are integration tests for networking aspects which however use only
local interfaces, and thus can be run unit-test style (but for example,
QEMU still needs to be run in networking mode, which requires special
daemons running). These tests are also generic socket tests in their
nature, but details of setup (addresses to use, ability to do such
testing at all) are specific to a particular environment, Zephyr in this
case, so for now made port-specific.